Two step adding process with selecting color
