+++ frontend/src/components/TagEditor.tsx
@@ -2,41 +2,53 @@ import React from 'react'
 import { MAX_NAME_LENGTH, REGEX_NAME_SAFE } from '../shared/constants'
 import { useSelector, useDispatch } from 'react-redux'
 import { ApplicationState, Dispatch } from '../store'
-import { InlineTextFieldSetting } from './InlineTextFieldSetting'
-import { attributeName } from '../shared/nameHelper'
-import { safeHostname } from '../shared/nameHelper'
-import { LabelButton } from '../buttons/LabelButton'
-import { getDevices } from '../models/accounts'
-import { makeStyles } from '@material-ui/core'
+import { makeStyles, Button } from '@material-ui/core'
+import { AutocompleteMenu } from './AutocompleteMenu'
+import { Icon } from './Icon'
 
 export const TagEditor: React.FC<{ device: IDevice }> = ({ device }) => {
   const css = useStyles()
-  const { devices } = useDispatch<Dispatch>()
-  const { hostname, nameBlacklist } = useSelector((state: ApplicationState) => ({
-    hostname: state.backend.environment.hostname,
-    nameBlacklist: getDevices(state)
-      .filter((device: IDevice) => !device.shared)
-      .map((d: IDevice) => d.name.toLowerCase()),
-  }))
+  const tags = useSelector((state: ApplicationState) => state.labels)
+
+  const [newValue, setNewValue] = React.useState<ILabel>()
+  const [open, setOpen] = React.useState<boolean>(false)
+  const buttonRef = React.useRef<HTMLButtonElement>(null)
+
+  const handleOpen = () => setOpen(!open)
+  const handleClose = () => setOpen(false)
+  const handleAddTag = tag => console.log('ADD TAG', tag)
 
   if (!device) return null
 
-  const name = attributeName(device)
-
   return (
     <>
-      <InlineTextFieldSetting
-        value={name}
-        label="Tag Editor"
-        disabled={device.shared}
-        maxLength={MAX_NAME_LENGTH}
-        filter={REGEX_NAME_SAFE}
-        onSave={name => {
-          devices.renameDevice({ ...device, name: name.toString() })
-          // if we use device attributes:
-          // device.attributes = { ...device.attributes, name: name.toString() }
-          // devices.setAttributes(device)
+      <Button variant="text" size="small" onClick={handleOpen} ref={buttonRef}>
+        <Icon name="plus" size="sm" inlineLeft />
+        add tag
+      </Button>
+      <AutocompleteMenu
+        items={tags}
+        open={open}
+        placeholder="New tag..."
+        targetEl={buttonRef.current}
+        onSelect={(action, tag) => {
+          if (action === 'new') setNewValue(tag)
+          else handleAddTag(tag)
         }}
+        onClose={handleClose}
+        allowAdding
+      />
+      <AutocompleteMenu
+        items={tags}
+        open={Boolean(newValue)}
+        placeholder="Choose color..."
+        targetEl={buttonRef.current}
+        onSelect={(action, tag) => {
+          tag = { ...(newValue || tag), color: tag.color }
+          setNewValue(undefined)
+          handleAddTag(tag)
+        }}
+        onClose={handleClose}
       />
     </>
   )
